+++ CHANGELOG.md
@@ -2,10 +2,11 @@
   * **Improved ext-glyph**.
      * [Added] support for ligature icons (e.g. [material icons](https://material.io/icons/)).
      * [Added] `icon` option can now return a dict to create a ligature icon.
-  * [Added] Material Design demo
   * [Added] #798 Tooltip support for icons (dynamic option `iconTooltip`).
   * [Improved] ext-table no longer needs empty tbody/tr if thead is present.
-  * [Added] Demo for jquery-confirm popup.
+  * [Fixed] #796 UMD requirements for node/CommonJS
+  * [Added] Material Design demo
+  * [Added] Demo for jquery-confirm popup
 
 # 2.26.0 / 2017-11-04
   * BREAKING: